LCMS-23 separates views, admin controllers have views in admin folder

+++ database/migrations/2016_02_09_132439_create_permission_user_table.php
@@ -17,7 +17,7 @@ class CreatePermissionUserTable extends Migration
             $table->integer('permission_id')->unsigned()->index();
             $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
             //$table->integer('user_id');
-            $table->integer('user_id')->unasigned()->index();
+            $table->unsignedBigInteger('user_id')->unasigned()->index();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
         });
